# Conflicts:
#	.vs/ggj17/v14/.suo
#	Assets/_Art/Character/Player.prefab
#	Assets/_Scripts/Player/PlayerController.cs
#	Assets/_Scripts/Player/PlayerWeaponDamageTrigger.cs
#	LKLog.txt

+++ Assets/_Scripts/Combat/ICombatEntity.cs
@@ -6,5 +6,5 @@ public interface ICombatEntity
 {
     GameObject gameObject { get; }
 
-    void InflictDamage( float p_damage );
+    void InflictDamage( int p_damage );
 }
